unique_everseen test refactor
